requirements.txt
django
django-admin-interface
django-crispy-forms
crispy-bootstrap4
pillow
zeep
psycopg2-binary
djangorestframework
django-admin-confirm
django-axes
django-simple-captcha
django-recaptcha
cloudinary
whitenoise

# pip install -r requirements.txt